+++ lua/obsidian/commands/footnotes.lua
@@ -1,0 +1,3 @@
+return function()
+  require("obsidian.footnotes").pick()
+end