+++ lua/obsidian/commands/init.lua
@@ -310,4 +310,6 @@ M.register("extract_note", { nargs = "?", range = true, note_action = true })
 
 M.register("toc", { nargs = 0, note_action = true })
 
+M.register("footnotes", { nargs = 0, note_action = true })
+
 return M