+++ tests/lsp/test_definition.lua
@@ -56,6 +56,40 @@ T["follow encoded headerlinks"] = function()
   eq(child.api.nvim_win_get_cursor(0), { 1, 0 })
 end
 
+T["goto footnote definition"] = function()
+  local files = h.mock_vault_contents(child.Obsidian.dir, {
+    ["note.md"] = [==[
+some claim[^1]
+
+more text
+
+[^1]: the footnote
+]==],
+  })
+  child.cmd("edit " .. files["note.md"])
+  child.api.nvim_win_set_cursor(0, { 1, 11 })
+  child.lua "vim.lsp.buf.definition()"
+  h.wait(function()
+    return vim.deep_equal(child.api.nvim_win_get_cursor(0), { 5, 0 })
+  end, { desc = "cursor on footnote definition" })
+end
+
+T["goto first footnote reference from definition"] = function()
+  local files = h.mock_vault_contents(child.Obsidian.dir, {
+    ["note.md"] = [==[
+some claim[^1]
+
+[^1]: the footnote
+]==],
+  })
+  child.cmd("edit " .. files["note.md"])
+  child.api.nvim_win_set_cursor(0, { 3, 0 })
+  child.lua "vim.lsp.buf.definition()"
+  h.wait(function()
+    return vim.deep_equal(child.api.nvim_win_get_cursor(0), { 1, 10 })
+  end, { desc = "cursor on first footnote reference" })
+end
+
 local filetypes = require("obsidian.attachment").filetypes
 
 local function test_ft(ext)